+++ domain/exposure-domain/build.gradle
@@ -9,8 +9,8 @@ dependencies {
     implementation project(":api-client:coupon-client")
     implementation project(":api-client:advertisement-client")
     implementation project(":domain:exposure-common-domain")
-//    implementation project(":domain:search-domain-rdb")
-    implementation project(":domain:search-domain-es")
+    implementation project(":domain:search-domain-rdb")
+    // implementation project(":domain:search-domain-es")
     implementation project(":domain:cache-domain-redis")
     implementation project(":event:event-publisher")
     implementation project(":event:common-event")
