--- mono/src/main/java/woowa/team4/bff/menu/category/controller/create/MenuCategoryCreateRequest.java
@@ -1,13 +1,0 @@
-package woowa.team4.bff.menu.category.controller.create;
-
-import jakarta.validation.constraints.NotBlank;
-import woowa.team4.bff.menu.category.command.MenuCategoryCreateCommand;
-
-public record MenuCategoryCreateRequest(
-        @NotBlank String name,
-        @NotBlank String description) {
-
-    public MenuCategoryCreateCommand toCommand(final String restaurantUuid) {
-        return new MenuCategoryCreateCommand(restaurantUuid, name, description);
-    }
-}
