--- mono/src/main/java/woowa/team4/bff/menu/category/controller/create/MenuCategoryCreateResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.category.controller.create;
-
-public record MenuCategoryCreateResponse(String menuCategoryUuid) {
-
-}
