--- mono/src/main/java/woowa/team4/bff/menu/category/controller/delete/MenuCategoryDeleteResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.category.controller.delete;
-
-public record MenuCategoryDeleteResponse(String menuCategoryUuid) {
-
-}
