--- mono/src/main/java/woowa/team4/bff/menu/category/controller/update/MenuCategoryUpdateResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.category.controller.update;
-
-public record MenuCategoryUpdateResponse(String menuCategoryUuid) {
-
-}
