--- mono/src/main/java/woowa/team4/bff/menu/item/controller/create/MenuCreateResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.item.controller.create;
-
-public record MenuCreateResponse(String menuUuid) {
-
-}
