--- mono/src/main/java/woowa/team4/bff/menu/item/controller/delete/MenuDeleteResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.item.controller.delete;
-
-public record MenuDeleteResponse(String menuUuid) {
-
-}
