--- mono/src/main/java/woowa/team4/bff/menu/item/controller/update/MenuUpdateResponse.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.menu.item.controller.update;
-
-public record MenuUpdateResponse(String menuUuid) {
-
-}
