--- mono/src/main/java/woowa/team4/bff/restaurant/controller/create/RegisterRestaurantResponse.java
@@ -1,4 +1,0 @@
-package woowa.team4.bff.restaurant.controller.create;
-
-public record RegisterRestaurantResponse(String restaurantId) {
-}
