--- mono/src/main/java/woowa/team4/bff/review/controller/create/ReviewMenuCreateRequest.java
@@ -1,9 +1,0 @@
-package woowa.team4.bff.review.controller.create;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record ReviewMenuCreateRequest(@NotBlank String menuUuid, @NotNull String content,
-                                      @NotNull Boolean isLiked) {
-
-}
