+++ service/build.gradle
@@ -1,0 +1,2 @@
+bootJar.enabled = false
+jar.enabled = false
