--- service/cache-service/src/main/java/woowa/team4/cacheservice/CacheServiceApplication.java
@@ -1,13 +1,0 @@
-package woowa.team4.cacheservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CacheServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CacheServiceApplication.class, args);
-    }
-
-}
