+++ service/cache-service/src/test/java/woowa/team4/bff/CacheServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package woowa.team4.bff;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CacheServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
