--- service/cache-service/src/test/java/woowa/team4/cacheservice/CacheServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.cacheservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CacheServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
