+++ service/coupon-service/src/main/resources/application.yml
