--- service/restaurant-exposure-service/src/test/java/woowa/team4/bff/restaurantexposureservice/RestauarntExposureServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.bff.restaurantexposureservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestauarntExposureServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
