+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/category/controller/create/MenuCategoryCreateResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.category.controller.create;
+
+public record MenuCategoryCreateResponse(String menuCategoryUuid) {
+
+}
