+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/category/controller/delete/MenuCategoryDeleteResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.category.controller.delete;
+
+public record MenuCategoryDeleteResponse(String menuCategoryUuid) {
+
+}
