+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/category/controller/update/MenuCategoryUpdateResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.category.controller.update;
+
+public record MenuCategoryUpdateResponse(String menuCategoryUuid) {
+
+}
