+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/item/controller/create/MenuCreateResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.item.controller.create;
+
+public record MenuCreateResponse(String menuUuid) {
+
+}
