+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/item/controller/delete/MenuDeleteResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.item.controller.delete;
+
+public record MenuDeleteResponse(String menuUuid) {
+
+}
