+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/menu/item/controller/update/MenuUpdateResponse.java
@@ -1,0 +1,5 @@
+package woowa.team4.bff.restaurantservice.menu.item.controller.update;
+
+public record MenuUpdateResponse(String menuUuid) {
+
+}
