+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/restaurant/controller/create/RegisterRestaurantResponse.java
@@ -1,0 +1,4 @@
+package woowa.team4.bff.restaurantservice.restaurant.controller.create;
+
+public record RegisterRestaurantResponse(String restaurantId) {
+}
