+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/review/controller/create/ReviewCreateResponse.java
@@ -1,0 +1,10 @@
+package woowa.team4.bff.restaurantservice.review.controller.create;
+
+import java.util.List;
+
+public record ReviewCreateResponse (String reviewId, List<String> reviewMenuIds) {
+
+        public static ReviewCreateResponse of(String reviewId, List<String> reviewMenuIds) {
+            return new ReviewCreateResponse(reviewId, reviewMenuIds);
+        }
+}
