+++ service/restaurant-service/src/main/java/woowa/team4/bff/restaurantservice/review/controller/create/ReviewMenuCreateRequest.java
@@ -1,0 +1,9 @@
+package woowa.team4.bff.restaurantservice.review.controller.create;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record ReviewMenuCreateRequest(@NotBlank String menuUuid, @NotNull String content,
+                                      @NotNull Boolean isLiked) {
+
+}
