+++ service/restaurant-service/src/test/java/woowa/team4/bff/RestauarntServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package woowa.team4.bff;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RestauarntServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
