--- service/restaurant-service/src/test/java/woowa/team4/restaurantservice/RestauarntServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.restaurantservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestauarntServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
