+++ service/search-service/src/test/java/woowa/team4/bff/SearchServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package woowa.team4.bff;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SearchServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
