--- service/search-service/src/test/java/woowa/team4/searchservice/SearchServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.searchservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SearchServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
