+++ settings.gradle
@@ -34,7 +34,20 @@ findProject(':domain:cache-domain-redis')?.name = 'cache-domain-redis'
 include 'dummy-api'
 include 'event:cache-event'
 findProject(':event:cache-event')?.name = 'cache-event'
-include 'external'
+include 'service:advertisement-service'
+findProject(':service:advertisement-service')?.name = 'advertisement-service'
+include 'service:coupon-service'
+findProject(':service:coupon-service')?.name = 'coupon-service'
+include 'service:delivery-time-service'
+findProject(':service:delivery-time-service')?.name = 'delivery-time-service'
+include 'service:restaurant-exposure-service'
+findProject(':service:restaurant-exposure-service')?.name = 'restaurant-exposure-service'
+include 'service:restaurant-service'
+findProject(':service:restaurant-service')?.name = 'restaurant-service'
+include 'service:cache-service'
+findProject(':service:cache-service')?.name = 'cache-service'
+include 'service:search-service'
+findProject(':service:search-service')?.name = 'search-service'
 include 'api-client'
 include 'api-client:common-client'
 findProject(':api-client:common-client')?.name = 'common-client'
